backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0
pydantic-settings>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.10
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-multipart>=0.0.7
pypdf>=4.0.0
python-docx>=1.1.0
reportlab>=4.0.0
